refactor flag tests
