+++ cmd/root.go
@@ -4,11 +4,8 @@ Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 package cmd
 
 import (
-	"errors"
-	"fmt"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	"github.com/spf13/viper"
 	"log"
 	"os"
 )
@@ -41,12 +38,12 @@ func Execute() {
 	}
 }
 
-func addGlobalPersistentFlags(f *pflag.FlagSet) {
-	f.StringVar(&cfgFile, "config", "", "config file (default is $HOME/.tykctl.yaml)")
-}
-
 func addRootLocalFlags(f *pflag.FlagSet) {
 	f.BoolP("toggle", "t", false, "Help message for toggle")
+}
+
+func addGlobalPersistentFlags(f *pflag.FlagSet) {
+	f.StringVar(&cfgFile, "config", "", "config file (default is $HOME/.tykctl.yaml)")
 }
 
 func init() {
@@ -55,50 +52,4 @@ func init() {
 		log.Fatal(err)
 	}
 	cobra.OnInitialize(initConfig)
-}
-
-// initConfig reads in config file and ENV variables if set.
-func initConfig() {
-	//cobra.CheckErr(err)
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := os.UserHomeDir()
-		cobra.CheckErr(err)
-
-		// Search config in home directory with name ".ara-client-sdk" (without extension).
-		viper.AddConfigPath(home)
-		viper.SetConfigType("yaml")
-		viper.SetConfigName(".tykctl")
-	}
-
-	viper.AutomaticEnv() // read in environment variables that match
-
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())
-	}
-
-}
-func createConfigFile() error {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-	result := fmt.Sprintf("%s/%s", home, ".tykctl.yaml")
-	_, err = os.Stat(result)
-	if errors.Is(err, os.ErrNotExist) {
-		f, err := os.Create(result)
-		log.Println(err)
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		return nil
-
-	}
-	return err
-
 }
