+++ cmd/root_test.go
@@ -1,0 +1,19 @@
+package cmd
+
+import (
+	"testing"
+)
+
+func TestNewRootCmd(t *testing.T) {
+	flags := []Flag{{
+		Description: "Test toggle flag",
+		Name:        "toggle",
+		Shorthand:   "t",
+		Default:     "false",
+		Value:       "false",
+	}}
+
+	cmd := NewRootCmd()
+	testFlags(t, cmd.Flags(), flags)
+
+}
